Make dependency versions less restrictive
Somewhat annoying, but because of how SEMVER works, there can be breaking changes between 0.x and 0.y versions (not between 1.x and 1.y), so that `KrylovKit = "0.4"` does not allow to use `v"0.5"`, leading to dependency conflicts. I've also added a few non-existing (yet) versions of my packages, to be somewhat future proof, including a version 1, which means this should be good to go up to version 2 (whenever that will ever happen).

Another food for thought: if you want less dependencies, you could remove OptimKit and TensorManifolds, and only make this ground state method available if these packages are separately loaded using the Require.jl mechanism. Not sure if that is worth the trouble though.